Refactor: move 'async_stream_actor_mock' helper function from test_stream_async to utils
